requirements.txt
Flask==3.0.0
google-generativeai==0.5.1
